+++ openrem/remapp/tests/test_dx_imports.py
@@ -3,9 +3,10 @@
 
 from __future__ import unicode_literals
 from django.test import TestCase
-from dicom.sequence import Sequence
+from dicom.dataelem import RawDataElement
 from dicom.dataset import Dataset
-from remapp.extractors.dx import _xray_filters_multiple
+from dicom.tag import Tag
+from remapp.extractors.dx import _xray_filters_prep
 from remapp.models import GeneralStudyModuleAttr, ProjectionXRayRadiationDose, IrradEventXRayData, \
     IrradEventXRaySourceData
 
@@ -14,10 +15,13 @@ class DXImportTests(TestCase):
         """
         Test the material extraction process when the materials are comma separated
         """
-        ds = Dataset()
-        FilterMaterial = "niobium,europium"
-        ds.FilterThicknessMinimum = "1.0\\0.1"
-        ds.FilterThicknessMaximum = "1.0\\0.1"
+
+        rawelemmin = RawDataElement(Tag(0x187052), 'DS', 9, '0.09,1.18', 0, False, True)
+        rawelemmax = RawDataElement(Tag(0x187054), 'DS', 9, '0.11,1.22', 0, False, True)
+        rawdict = {0x187052: rawelemmin, 0x187054: rawelemmax}
+        ds = Dataset(rawdict)
+
+        ds.FilterMaterial = "niobium,europium"
 
         g = GeneralStudyModuleAttr.objects.create()
         g.save()
@@ -28,9 +32,12 @@ class DXImportTests(TestCase):
         source = IrradEventXRaySourceData.objects.create(irradiation_event_xray_data=event)
         source.save()
 
-        _xray_filters_multiple(FilterMaterial, ds.FilterThicknessMaximum, ds.FilterThicknessMinimum, source)
+        _xray_filters_prep(ds, source)
 
-        self.assertEqual(source.xrayfilters_set.all().count(), 2)
+        self.assertEqual(source.xrayfilters_set.all().count(), 2,
+                         "Testing Kodak old style, two filters should have been stored, {0} were".format(
+                             source.xrayfilters_set.all().count()
+                         ))
         self.assertEqual(source.xrayfilters_set.all()[0].xray_filter_material.code_meaning,
                          "Niobium or Niobium compound")
         self.assertEqual(source.xrayfilters_set.all()[1].xray_filter_material.code_meaning,
@@ -55,7 +62,7 @@ class DXImportTests(TestCase):
         source = IrradEventXRaySourceData.objects.create(irradiation_event_xray_data=event)
         source.save()
 
-        _xray_filters_multiple(ds.FilterMaterial, ds.FilterThicknessMaximum, ds.FilterThicknessMinimum, source)
+        _xray_filters_prep(ds, source)
 
         self.assertEqual(source.xrayfilters_set.all().count(), 2, 'Wrong number of filters recorded')
         self.assertEqual(source.xrayfilters_set.all()[0].xray_filter_material.code_meaning,
@@ -64,27 +71,26 @@ class DXImportTests(TestCase):
                          "Copper or Copper compound")
 
 
-    # def test_single_filter(self):
-    #     """
-    #     Test the material extraction process when there is just one filter
-    #     *** Tests the wrong bit of code, so commenting out
-    #     """
-    #     ds = Dataset()
-    #     FilterMaterial = "lead"
-    #     ds.FilterThicknessMinimum = "1.0"
-    #     ds.FilterThicknessMaximum = "1.0"
-    #
-    #     g = GeneralStudyModuleAttr.objects.create()
-    #     g.save()
-    #     proj = ProjectionXRayRadiationDose.objects.create(general_study_module_attributes=g)
-    #     proj.save()
-    #     event = IrradEventXRayData.objects.create(projection_xray_radiation_dose=proj)
-    #     event.save()
-    #     source = IrradEventXRaySourceData.objects.create(irradiation_event_xray_data=event)
-    #     source.save()
-    #
-    #     _xray_filters_multiple(FilterMaterial, ds.FilterThicknessMaximum, ds.FilterThicknessMinimum, source)
-    #
-    #     self.assertEqual(source.xrayfilters_set.all().count(), 1)
-    #     self.assertEqual(source.xrayfilters_set.all()[0].xray_filter_material.code_meaning,
-    #                      "Lead or Lead compound")
+    def test_single_filter(self):
+        """
+        Test the material extraction process when there is just one filter
+        """
+        ds = Dataset()
+        ds.FilterMaterial = "lead"
+        ds.FilterThicknessMinimum = "1.0"
+        ds.FilterThicknessMaximum = "1.0"
+
+        g = GeneralStudyModuleAttr.objects.create()
+        g.save()
+        proj = ProjectionXRayRadiationDose.objects.create(general_study_module_attributes=g)
+        proj.save()
+        event = IrradEventXRayData.objects.create(projection_xray_radiation_dose=proj)
+        event.save()
+        source = IrradEventXRaySourceData.objects.create(irradiation_event_xray_data=event)
+        source.save()
+
+        _xray_filters_prep(ds, source)
+
+        self.assertEqual(source.xrayfilters_set.all().count(), 1)
+        self.assertEqual(source.xrayfilters_set.all()[0].xray_filter_material.code_meaning,
+                         "Lead or Lead compound")