+++ docs/changes.rst
@@ -5,7 +5,7 @@ OpenREM version history
 0.8.0 (?)
 ---------
 
-* `#450`_  Imports: X-ray filter thicknesses of 0.00 now recorded as such
+* `#450`_  Imports: DX imports with filter thickness of 0.00 are now recorded as such
 * `#443`_  Exports: Accession number and ID are now exported to XLSX as text. Thanks to `@LuukO`_
 * `#441`_  Charts: Fixed a bug that broke chart links containing non-ASCII characters
 * `#440`_  Charts: Fixed a bug in sorting.js so that undefined strings are handled correctly